Don't apply clipping for the root element.
The idea is that the root element coincides with the canvas, so anything
outside the bounds of the root element would be outside the canvas
anyway. This isn't quite true in the fancy of fancy CSS transforms, but
it's true in normal cases. Mainly, it works around a bug in html2canvas
where parent clipping regions are applied incorrectly when rendering
child elements. In TerriaJS, this manifests as point markers (e.g. in
`#test`'s TerriaJS Test Data -> GeoJSON -> Test overriding styled and
unstyled features) not showing up in the generated image. More details
of the html2canvas issue can be found here:
https://github.com/niklasvh/html2canvas/pull/1084